refactor: Improve dependency hierarchy between Core, Model and Platform layers
- Add IBaseItem interface in Core layer
- Update IMTTreeItem to inherit from IBaseItem
- Fix import paths to maintain proper dependency direction
- Adjust references across interfaces to align with layered architecture
