+++ temp/core/base_item_interface.py
@@ -1,0 +1,9 @@
+from typing import Protocol, Dict, Any
+
+class IBaseItem(Protocol):
+    """기본 아이템 인터페이스"""
+    @property
+    def id(self) -> str: ...
+    
+    @property
+    def data(self) -> Dict[str, Any]: ...
