+++ temp/model/tree_item_interface.py
@@ -1,5 +1,6 @@
 from typing import Protocol, List, Optional, Dict, Any, Tuple, Union, TypedDict, TypeVar, Generic
 from enum import Enum
+from temp.core.base_item_interface import IBaseItem  # 새로 추가
 
 
 class MTNode(Enum):
@@ -89,21 +90,13 @@ class TreeItemData(TypedDict, total=False):
 
 T = TypeVar('T')  # 제네릭 타입 변수 정의
 
-class IMTTreeItem(Protocol):
+class IMTTreeItem(IBaseItem, Protocol):  # IBaseItem 상속 추가
     """매크로 트리 아이템 인터페이스
     
     트리에서 사용되는 개별 아이템의 인터페이스입니다.
     id만 직접 접근 속성으로 제공하고, 나머지는 data 딕셔너리를 통해 액세스합니다.
     """
-    @property
-    def id(self) -> str:
-        """아이템 고유 식별자"""
-        ...
-    
-    @property
-    def data(self) -> TreeItemData:
-        """아이템의 모든 데이터"""
-        ...
+    # id와 data 프로퍼티는 IBaseItem에서 이미 정의됨
     
     def set_property(self, key: str, value: T) -> None:
         """아이템 속성을 설정합니다."""
